+++ src/promptbase/gsm8k/gsm8k.py
@@ -2,7 +2,7 @@
 import json
 import pathlib
 
-from promptbase.utils import text_completion, run_batch_jobs
+from promptbase.utils.helpers import text_completion, run_batch_jobs
 from datasets import load_dataset
