+++ src/promptbase/utils/__init__.py
@@ -1,1 +1,2 @@
-from .utils import text_completion, run_batch_jobs, fetch_dataset_blob
+from . import helpers
+from . import consts
